refactor: изменить логику применения стилей
