+++ src/components/article-maintext/article-maintext.stories.tsx
@@ -2,42 +2,7 @@ import {ComponentMeta, ComponentStory} from '@storybook/react';
 
 import { ArticleMainText } from './article-maintext';
 
-function ArticleText() {
-  return (
-    <>
-      <h6>В образовательной программе «Драматург плюс стратегии и практики»прошел мастер-класс
-      «Документальность в современном театре и современном искусстве». Его провели драматург,
-      сценарист, преподаватель Евгений Казачков и драматург, художница, куратор, кандидат филологических
-      наук Ильмира Болотян. Мы приводим расшифровку мастер-класса и обсуждения, развернувшегося после него.</h6>
-
-      <h4>Что нам говорил Аристотель? Что искусство должно быть подобно реальности.</h4>
-
-      <p>Из этого следует два вывода: с одной стороны, искусство не должно быть равно реальности,
-    потому что иначе оно не будет искусством, с другой стороны – оно должно быть все-таки подобно реальности,
-    иначе мы ничего не поймем, не подключимся и не распознаем замыслы автора. Мне нравится такое определение,
-    что один из видов эстетики – это, грубо говоря, понимание чего-то. Если нам нечего понимать, то мы не
-    получим ни удовольствия, ни какого-то чувства, ничего. А для того, чтобы что-то понять, нам должно быть
-    представлено что-то, что не равно само себе. То есть нам показывают или говорят что-то, а мы понимаем, что
-    там есть что-то еще. Вот мы смотрим на античную статую и понимаем, что это – статуя, но похожа на человека.
-    Если она будет абсолютно такая же как человек, как восковая фигура, то мы поймем, что это уже не искусство.
-    Искусство для того, чтобы люди догадались, домыслили, дочувствовали что-то.</p>
-
-      <blockquote>Настал именно режиссерский театр, потому что пространство высказывания оказалось
-    все в руках режиссеров, и они использовали доступную драматургию, классическую или импортную,
-    чтобы сделать свое высказывание о реальности.</blockquote>
-
-      <p>Из этого следует два вывода: с одной стороны, искусство не должно быть равно реальности,
-        потому что иначе оно не будет искусством, с другой стороны – оно должно быть все-таки подобно реальности,
-        иначе мы ничего не поймем, не подключимся и не распознаем замыслы автора. Мне нравится такое определение,
-        что один из видов эстетики – это, грубо говоря, понимание чего-то. Если нам нечего понимать, то мы не
-        получим ни удовольствия, ни какого-то чувства, ничего. А для того, чтобы что-то понять, нам должно быть
-        представлено что-то, что не равно само себе. То есть нам показывают или говорят что-то, а мы понимаем, что
-        там есть что-то еще. Вот мы смотрим на античную статую и понимаем, что это – статуя, но похожа на человека.
-        Если она будет абсолютно такая же как человек, как восковая фигура, то мы поймем, что это уже не искусство.
-        Искусство для того, чтобы люди догадались, домыслили, дочувствовали что-то.</p>
-    </>
-  );
-}
+import ArticleText from './assets/mock-data-articleText';
 
 export default {
   title: 'components/ArticleMainText',
